# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/powerbanks": {
            "get": {
                "summary": "Get info about all powerbanks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Powerbank"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create powerbank",
                "parameters": [
                    {
                        "description": "Serial number of powerbank",
                        "name": "SerialNumber",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Full capacity on powerbank",
                        "name": "Capacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "Status on use's powerbanks",
                        "name": "Used",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Powerbank"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/powerbanks/{powerbankId}": {
            "get": {
                "summary": "Get info about powerbank",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank Id",
                        "name": "powerbankId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Powerbank"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Create powerbank",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank Id",
                        "name": "powerbankId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status on use's powerbanks",
                        "name": "Used",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Powerbank"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Create powerbank",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank Id",
                        "name": "powerbankId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Powerbank": {
            "type": "object",
            "properties": {
                "capacity": {
                    "description": "объем заряда",
                    "type": "number"
                },
                "create_at": {
                    "type": "string"
                },
                "delete_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "removed": {
                    "type": "boolean"
                },
                "serial_number": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "used": {
                    "description": "сколько уже использована банка в часах // up юзается или нет",
                    "type": "boolean"
                }
            }
        },
...
    }
}